refactor: replace useMetrics with useAnalytics in Bridge components
